shared/pyproject.toml
[tool.poetry]
name = "shared"
version = "0.1.0"
description = ""
authors = ["AshisGhosh <ashisghosh@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
transformers = "^4.38.1"
torch = "^2.2.1+cpu"
torchvision = "^0.17.1+cpu"
torchaudio = "^2.2.1+cpu"
httpx = ">=0.15.4,<0.26.0"
python-dotenv = "^1.0.1"
langfuse = "^2.16.2"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
